+++ src/components/LandingPage/AboutMentorsSection.tsx
@@ -2,34 +2,29 @@ import Button from "@components/Button";
 import Image from "next/image";
 import Link from "next/link";
 
-type Props = {};
-
-const AboutMentorsSection = (props: Props) => {
+const AboutMentorsSection = () => {
   return (
-    <section
-      className="bg-gray-01 dark:bg-secondary-02 py-[70px]"
-      id="mentores"
-    >
-      <div className="container flex justify-between items-center">
-        <div>
-          <h3 className="text-gray-04 dark:text-gray-01 text-xl font-bold">
+    <section className="bg-secondary-02 py-[70px] px-2 sm:px-0" id="mentores">
+      <div className="xs:container flex flex-col lg:flex-row justify-between items-center">
+        <div className="flex flex-col ">
+          <h3 className="text-gray-01 text-xl font-bold text-center lg:text-start">
             SOBRE NOSSOS MENTORES
           </h3>
-          <h2 className="text-[2rem] font-bold max-w-[294px] leading-[1.0] mt-6">
+          <h2 className="text-center lg:text-start text-2xl xs:text-[2rem] font-bold sm:max-w-[294px] leading-[1.0] mt-6 text-neutral-01 m-auto lg:m-0 lg:py-4">
             Conheça todos os nossos mentores
           </h2>
-          <p className="max-w-[365px] mt-4 text-gray-04">
+          <p className="max-w-[365px] mt-4 text-gray-01 text-center lg:text-start px-4 xs:px-0 m-auto">
             Profissionais incríveis com experiênciais reais de mercado para
             compartilhar com você. Converse com mentores da área da tecnologia e
             prepare-se para os desafios.
           </p>
-          <div className="mt-12  max-w-[220px]">
-            <Link href={"/"}>
+          <div className="mt-12  max-w-[220px] m-auto lg:m-0 lg:py-12 w-full">
+            <Link href={"/signin"} className="focus:outline-none" tabIndex={1}>
               <Button size="small">Saiba mais</Button>
             </Link>
           </div>
         </div>
-        <div>
+        <div className="mt-10 lg:m-0 px-4 xs:px-0 max-w-[500px] 2xl:max-w-max">
           <Image
             alt="mentor learning"
             src={"/mentor-learning.png"}
