+++ src/components/LandingPage/CardInstructionSection.tsx
@@ -12,9 +12,11 @@ const CardInstructionSection: React.FC<CardType> = ({
   title,
 }) => {
   return (
-    <div className=" bg-secondary-02 rounded-lg px-14 py-12 max-w-[350px]">
-      <Icon size={32} className="text-neutral-01" />
-      <h3 className="font-bold text-4.5xl text-neutral-02">{title}</h3>
+    <div className=" bg-secondary-02 rounded-lg py-4 px-8 xs:py-8 xs:px-12 max-w-[350px] space-y-2">
+      <Icon size={32} className="text-neutral-01 mb-2 sm:mb-0" />
+      <h3 className="font-bold text-lg sm:text-xl md:text-2xl lg:text-3xl xl:text-4xl text-neutral-02">
+        {title}
+      </h3>
       <p className="text-neutral-05 dark:text-neutral-01">{text}</p>
     </div>
   );
