+++ src/components/LandingPage/Footer/About.tsx
@@ -1,17 +1,25 @@
-import React from "react";
-
 const About = () => {
   return (
-    <section>
-      <h4 className="text-2xl font-bold">Sobre nós</h4>
-      <ul className="mt-6 space-y-2">
-        <li className="hover:opacity-70 hover:cursor-pointer">Notion</li>
-        <li className="hover:opacity-70 hover:cursor-pointer">
-          Repositório Back End
-        </li>
-        <li className="hover:opacity-70 hover:cursor-pointer">
-          Repositório Front End
-        </li>
+    <section className="py-4 lg:py-0">
+      <h4 className="text-2xl font-bold text-neutral-01 text-center lg:text-start">
+        Sobre nós
+      </h4>
+      <ul className="mt-6 flex flex-col gap-2 justify-center items-center lg:items-start lg:justify-start mb-10 xs:mb-10">
+        <a href="https://uxdavidareal.notion.site/Mentor-Cycle-ff5e6e73778d40d690d57aeac24d8cb5">
+          <li className="hover:opacity-70 hover:cursor-pointer text-gray-01">
+            Notion
+          </li>
+        </a>
+        <a href="https://github.com/Mentor-Cycle/mentor-cycle-fe">
+          <li className="hover:opacity-70 hover:cursor-pointer text-gray-01">
+            Repositório Back End
+          </li>
+        </a>
+        <a href="https://github.com/Mentor-Cycle/mentor-cycle-be">
+          <li className="hover:opacity-70 hover:cursor-pointer text-gray-01">
+            Repositório Front End
+          </li>
+        </a>
       </ul>
     </section>
   );
