+++ src/components/LandingPage/Footer/Contact.tsx
@@ -1,8 +1,14 @@
 const Contact = () => {
   return (
-    <section>
-      <h4 className="text-2xl font-bold mb-6">Contatos</h4>
-      <span>https://www.linkedin.com/company/mentor-cycle</span>
+    <section className="flex flex-col xs justify-center xs:justify-start items-center">
+      <h4 className="text-2xl font-bold mb-6 text-neutral-01 text-center xs:text-start">
+        Contatos
+      </h4>
+      <a href="https://www.linkedin.com/company/mentor-cycle">
+        <span className="text-gray-01 hover:text-gray-02">
+          Organização no Linkedin
+        </span>
+      </a>
     </section>
   );
 };
