+++ src/components/LandingPage/Footer/SocialMediasSection.tsx
@@ -1,39 +1,36 @@
 import { useUser } from "@hooks/useUser";
-import dynamic from "next/dynamic";
+import Image from "next/image";
 import Link from "next/link";
-import { AiOutlineTwitter } from "react-icons/ai";
 import { BsDiscord } from "react-icons/bs";
 import { TiSocialLinkedinCircular } from "react-icons/ti";
 
 const SocialMediasSection = () => {
   const { user } = useUser();
-  const DynamicThemedImage = dynamic(
-    () => import("@components/Header/ThemeImage"),
-    {
-      ssr: false,
-    }
-  );
 
-  const userIsLogged = user.isLogged ? "/dashboard" : "/";
+  const userIsLogged = user.isLogged ? "/dashboard" : "/signin";
   return (
     <section>
-      <div className="hidden sm:flex justify-start items-center">
+      <div className="flex justify-center lg:justify-start items-center">
         <Link href={userIsLogged}>
-          <DynamicThemedImage />
+          <Image alt="Logo" src={"/logoDarkMode.png"} width={55} height={55} />
         </Link>
       </div>
-      <p className="max-w-[238px] mt-6 text-secondary-02">
+      <p className="lg:max-w-[238px] mt-6 text-neutral-01 text-center lg:text-start">
         Mentor Cycle - conectando mentes em busca de sucesso.
       </p>
-      <div className="flex mt-8 gap-[38px] justify-start items-center">
-        <BsDiscord
-          size={24}
-          className="hover:opacity-70 hover:cursor-pointer"
-        />
-        <TiSocialLinkedinCircular
-          size={30}
-          className="hover:opacity-70 hover:cursor-pointer"
-        />
+      <div className="flex mt-8 gap-[38px] justify-center mb-10 xs:mb-0 lg:justify-start items-center">
+        <a href={"https://discord.gg/WRD3uT3JaC"}>
+          <BsDiscord
+            size={24}
+            className="hover:opacity-70 hover:cursor-pointer text-neutral-02"
+          />
+        </a>
+        <a href="https://www.linkedin.com/company/mentor-cycle/mycompany/">
+          <TiSocialLinkedinCircular
+            size={30}
+            className="hover:opacity-70 hover:cursor-pointer text-neutral-02"
+          />
+        </a>
       </div>
     </section>
   );
