+++ src/components/LandingPage/Header/scrollSection.ts
@@ -1,10 +1,7 @@
-import { HEIGHT_HEADER } from "./HeaderLP";
-
 export const scrollToSection = (id: string) => {
   const section = document.getElementById(id);
   if (section) {
-    const sectionTopOffset =
-      section.getBoundingClientRect().top + window.scrollY - HEIGHT_HEADER;
-    window.scrollTo({ top: sectionTopOffset, behavior: "smooth" });
+    const y = section.getBoundingClientRect().top + window.scrollY - 84;
+    window.scrollTo({ top: y, behavior: "smooth" });
   }
 };
